requirements.txt
fastapi
uvicorn
motor
python-dotenv
pydantic
python-jose[cryptography]
passlib[bcrypt]
python-multipart
faker